Update products database migration
After using scaffold to prototype a model, it is helpful to edit the database migration to add database constraints that allow the PostgreSQL database to help enforce data integrity. This informs PostgreSQL the following:

- name, published, and category must be specified and cannot be blank.
- published is defaulted to false

These constraints are obvious from the product design. We do not have to wait until later to add them just because we are practicing TDD.