Updated model to use validation rather than enum for the category
I suggest students avoid using enum for this excercise. The string value set by the form was not being automatically recognized and the finess needed to make this work is beyond the scope of this excercise. Enums are a good tool and should be evaluated for use in real applications.