Add failing test for the home page showing the count
This update adds a failing test because the home page does not have a link to the products page with a count included in the description.

I've chosen to demonstrate how FactoryBot and Faker can be used in combination to provide lists of test data in the database. The rspec test
simply asks for 5 published products and 3 unpublished products and then can check the homepage to see that it references the correct count
as part of the anchor text for a link.