Adds a cookie-based Shopping Card
This implements a spike for a cookie-based shopping cart with add and remove functionality. The cart is maintaned by a plain Ruby object that is not backed by a database table. The contents are persisted as a hash stored in a JSON value in a cookie. Rails encrypts the cookie by default and protects it from forgery.