Add OrderProduct model
This is the object that represents a line item on an order. It serves as a mapping table between Order and Product and holds meta data associated with that relationship, like a prescription number for prescription products.

This update is the result of running:
bundle exec rails g model OrderProduct order:references product:references prescription_number:string

And then running
bundle exec rails db:migrate