Add new order page
This starts the minimum neccesary to load a new order page. It instantiates an order object with the current user (which is nil if no user is signed in) and authorizes it with a Pundit policy. It then renders a new page to which we will be adding a form.